Add meta-virtualization layer
meta-virtualization is a layer to enabling hypervisor, virtualization
tool stack, and cloud support.

meta-virtualization is hosted by Yocto Project and we need to create
a remote to fetch sources from git.yoctoproject.org. This remote is
called yocto and can be used to fetch any sources from git.yoctoproject.org

Change-Id: If6e97c13d7ba3bf83b54d367eea60ae6e87b40da
Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>